# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WhistleDrop",
    "description": "\nA confidential reporting API. Anyone can submit a report without an account or a name,\nand track it later with the case code they get back.\n\n**Reporters:** submit with `POST /api/reports` and save the `case_code` from the response.\nTo check on it, call `GET /api/reports/status` with the code in the `X-Case-Code` header.\n\n**Moderators:** log in with `POST /api/auth/login`, click **Authorize** and paste the `access_token`.\n\nEvery error looks like `{\"error\": {\"code\", \"message\", \"details\"}}`.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/reports": {
      "post": {
        "tags": [
          "Reporters"
        ],
        "summary": "Submit a report",
        "description": "Submit a report. No account and no name needed. The response holds your case code, shown only once.",
        "operationId": "submit_report_api_reports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitOut"
                }
              }
            }
          },
          "400": {
            "description": "The body isn't valid JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "413": {
            "description": "The body is over 32 KB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "A field failed validation, or an unknown field was sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "More than 10 reports in an hour from the same client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/moderator/reports": {
      "get": {
        "tags": [
          "Moderators"
        ],
        "summary": "List and filter reports",
        "description": "List reports, newest first by default. All filters are optional and can be combined.",
        "operationId": "list_reports_api_moderator_reports_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Status"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Category"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive text search in the description.",
              "title": "Q"
            },
            "description": "Case-insensitive text search in the description."
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Submitted on or after this date (YYYY-MM-DD).",
              "title": "From"
            },
            "description": "Submitted on or after this date (YYYY-MM-DD)."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Submitted on or before this date (YYYY-MM-DD).",
              "title": "To"
            },
            "description": "Submitted on or before this date (YYYY-MM-DD)."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "newest",
                "oldest"
              ],
              "type": "string",
              "default": "newest",
              "title": "Sort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportPage"
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is invalid or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "A filter has a bad value, or an unknown filter was sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/moderator/reports/{report_id}": {
      "get": {
        "tags": [
          "Moderators"
        ],
        "summary": "Get one report",
        "description": "One report in full, with every update, including internal notes.",
        "operationId": "get_report_api_moderator_reports__report_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetail"
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is invalid or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "No report has that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "The id isn't a UUID, or a field failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Moderators"
        ],
        "summary": "Change a report's status",
        "description": "Move a report to its next status: SUBMITTED to UNDER_REVIEW, then UNDER_REVIEW to RESOLVED or DISMISSED.",
        "operationId": "change_status_api_moderator_reports__report_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusChangeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetail"
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is invalid or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "No report has that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "The id isn't a UUID, or a field failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "The move isn't allowed, the case is closed, or someone else changed it first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/moderator/reports/{report_id}/updates": {
      "post": {
        "tags": [
          "Moderators"
        ],
        "summary": "Add a note",
        "description": "Add a note without changing the status. Set visible_to_reporter to false for an internal note.",
        "operationId": "add_update_api_moderator_reports__report_id__updates_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModeratorUpdateOut"
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is invalid or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "No report has that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "The id isn't a UUID, or a field failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "The case is closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/moderator/reports/{report_id}/close": {
      "post": {
        "tags": [
          "Moderators"
        ],
        "summary": "Close a case for good",
        "description": "Permanently close a RESOLVED or DISMISSED report. After this, nothing on it can change.",
        "operationId": "close_report_api_moderator_reports__report_id__close_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetail"
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is invalid or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "No report has that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "The id isn't a UUID, or a field failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "The report isn't RESOLVED or DISMISSED yet, or is already closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/moderator/stats": {
      "get": {
        "tags": [
          "Moderators"
        ],
        "summary": "Dashboard counts",
        "description": "How many reports there are by status and by category, and how many still need attention.",
        "operationId": "stats_api_moderator_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOut"
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is invalid or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/moderator/reports/{report_id}/evidence/{evidence_id}": {
      "get": {
        "tags": [
          "Moderators"
        ],
        "summary": "Download an evidence file",
        "description": "Download a file the reporter attached. It's always sent as a download, never shown inline.",
        "operationId": "download_evidence_api_moderator_reports__report_id__evidence__evidence_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          },
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The file",
            "content": {
              "image/jpeg": {},
              "image/png": {},
              "application/pdf": {}
            }
          },
          "401": {
            "description": "No token, or the token is invalid or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "No report has that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "The id isn't a UUID, or a field failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "string",
        "enum": [
          "SECURITY",
          "HARASSMENT",
          "CORRUPTION",
          "TECHNICAL",
          "OTHER"
        ],
        "title": "Category"
      },
...
      "ErrorOut": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorOut"
      },
...
      "ReportIn": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 20,
            "title": "Description"
          },
          "evidence_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "category",
          "description"
        ],
        "title": "ReportIn",
        "examples": [
          {
            "category": "SECURITY",
            "description": "The server room in the Tech Park basement is left unlocked every night after 9pm.",
            "evidence_url": "https://drive.google.com/file/d/1AbCdEf/view"
          }
        ]
      },
...
      "ReportPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReportListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "ReportPage"
      },
...
      "Status": {
        "type": "string",
        "enum": [
          "SUBMITTED",
          "UNDER_REVIEW",
          "RESOLVED",
          "DISMISSED"
        ],
        "title": "Status"
      },
...
      "SubmitOut": {
        "properties": {
          "case_code": {
            "type": "string",
            "title": "Case Code"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "submitted_on": {
            "type": "string",
            "format": "date",
            "title": "Submitted On"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "privacy_tip": {
            "type": "string",
            "title": "Privacy Tip"
          }
        },
        "type": "object",
        "required": [
          "case_code",
          "status",
          "submitted_on",
          "note",
          "privacy_tip"
        ],
        "title": "SubmitOut",
        "examples": [
          {
            "case_code": "WD-7K3M-Q9XA-2HFD-R8TN",
            "note": "Save this code now. It's the only way to check on your report and it can't be recovered.",
            "privacy_tip": "We don't know who you are, but details in your description that only you would know can still point back to you.",
            "status": "SUBMITTED",
            "submitted_on": "2026-10-04"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Paste the access_token from POST /api/auth/login.",
        "scheme": "bearer"
      }
    }
  },
...
}